made changes recommended by unit test but failing
I made changes recommended and left with just two that are still
coming back with an error, namely the `issued` and `modified` date
values on the edition file. i've changed them to pass the regex test
but for some reason still says failed, and says the old values are
still being entered e.g. "2023-05-05T08:59:11.430214" for issued.